+++ Source/Core/Utils/Allocator.h
@@ -1,10 +1,12 @@
+#include "Core/Core/BasicTypes.h"
+
 #include <new.h>
 
 class Allocator
 {
 public:
-	static void* Alloc(size_t size);
-	static void* Realloc(void* block, size_t size);
+	static void* Alloc(UInt64 size);
+	static void* Realloc(void* block, UInt64 size);
 	static void Free(void* block);
 
 	template<typename T, typename... Args>
